+++ Backend/controllers/gif.js
@@ -1,11 +1,8 @@
 const { Gif } = require("../sequelize").models;
 const fs = require('fs');
-const jwt = require('jsonwebtoken');
+
 
 exports.createGif = (req, res, next) => {
-    const token = req.headers.authorization.split(' ')[1];
-    const decodedToken = jwt.verify(token, 'RANDOM_TOKEN_SECRET');
-    const userId = decodedToken.userId;
     Gif.create({
         title: req.body.title,
         imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`,
@@ -13,11 +10,11 @@ exports.createGif = (req, res, next) => {
         dislikes : 0,
         usersLiked : "",
         usersDisliked : "",
-        userId: userId,
+        userId: req.userId,
 
     })
     .then(() => res.status(201).json({ message: 'Gif enregistré !'}))
-    .catch(error => res.status(400).json({ message: 'ça marche paaaaas !' }));
+    .catch(error => res.status(400).json({ message: 'ça marche pas !' }));
 };
 
 exports.getAllGif = (req, res, next) => {
@@ -36,12 +33,17 @@ exports.getOneGif = (req, res, next) => {
 exports.deleteGif = (req, res, next) => {
     Gif.findOne({where: { id: req.params.id }})
     .then(gif => {
-        const filename = gif.imageUrl.split('/images/')[1];
-        fs.unlink(`images/${filename}`, () => {
-            Gif.destroy( {where:{ id: req.params.id }})
-            .then(() => res.status(200).json({ message: 'Gif supprimé !'}))
-            .catch(error => res.status(400).json({ error }));
-        });
+        if (req.isAdmin || req.userId == gif.userId) {
+            const filename = gif.imageUrl.split('/images/')[1];
+            fs.unlink(`images/${filename}`, () => {
+                Gif.destroy( {where:{ id: req.params.id }})
+                .then(() => res.status(200).json({ message: 'Gif supprimé !'}))
+                .catch(error => res.status(400).json({ error }));
+            });
+        }
+        else {
+            return res.status(401).json ({error: 'pas autorisé'})
+        }
     })
     .catch(error => res.status(500).json({ error }));
 };
