+++ Backend/controllers/user.js
@@ -57,18 +57,27 @@ exports.getOneUser = (req, res, next) => {
 
 exports.deleteUser = (req, res, next) => {
     User.findOne({where:{ id: req.params.id }})
-    .then(c => {
-        
-        User.destroy({where:{ id: req.params.id }})
-        .then(() => res.status(200).json({ message: 'Commentaire supprimé !'}))
-        .catch(error => res.status(400).json({ error }));
-   
+    .then(user => {
+        if (req.isAdmin || req.userId == user.id) {        
+            User.destroy({where:{ id: req.params.id }})
+            .then(() => res.status(200).json({ message: 'Commentaire supprimé !'}))
+            .catch(error => res.status(400).json({ error }));
+    }
+    else {
+        return res.status(401).json ({error: 'pas autorisé'})
+    }
     })
     .catch(error => res.status(500).json({ error }));
 }; 
 
 exports.getAllUser = (req, res, next) => {
-    User.findAll()
-    .then(users => res.status(200).json(users))
-    .catch(error => res.status(400).json({ error }));
+    if (req.isAdmin ) {
+        User.findAll()
+        .then(users => res.status(200).json(users))
+        .catch(error => res.status(400).json({ error }));
+
+    }
+    else {
+        return res.status(401).json ({error: 'requête non autorisé'})
+    }
 };
